Break backup logic into `CROSSWORD--READ-PUZ-EMACS`.
The `CROSSWORD--SUMMARY-DATA-PUZ-EMACS` and `CROSSWORD-RESTORE` both
have similar logic.  Split it into a helper function both can use.

Also, fix `CROSSWORD--INSERT-CLUES` so it doesn’t stomp on the buffer
state.  This allows `CROSSWORD-RESTORE` to dump a bunch of `LET`
bindings for buffer-locals that otherwise become invisible due to the
buffer changing out from under it.